requirements.txt
matplotlib~=3.9.3
pillow~=11.0.0
opencv-python~=4.10.0.84
numpy~=2.1.3
pandas~=2.2.3